@@ -23,58 +23,66 @@ license = "ISC"
 readme = "README.md"
 repository = "https://github.com/starship/starship"
 # Note: MSRV is only intended as a hint, and only the latest version is officially supported in starship.
-rust-version = "1.60"
+rust-version = "1.64"
 description = """
 The minimal, blazing-fast, and infinitely customizable prompt for any shell! ☄🌌️
 """
 
 [features]
-default = ["battery", "notify"]
+default = ["battery", "notify", "git-repository-max-perf"]
 battery = ["starship-battery"]
 config-schema = ["schemars"]
 notify = ["notify-rust"]
 
+# Enables most of the `max-performace` features of the `git_repository` module for better performance.
+# This can be more difficult to build in some conditions and requires cmake.
+git-repository-max-perf = ["git-features/zlib-ng", "git-repository/fast-sha1"]
+# Slower than `git-repository-max-perf`, but better than the default.
+# Unlike `git-repository-max-perf` this does not require cmake and allows dynamic zlib linking.
+git-repository-faster = ["git-features/zlib-stock", "git-repository/fast-sha1"]
+
 [dependencies]
-chrono = { version = "0.4.22", features = ["clock", "std"] }
-clap = { version = "4.0.4", features = ["derive", "cargo", "unicode"] }
-clap_complete = "4.0.2"
+chrono = { version = "0.4.23", features = ["clock", "std"] }
+clap = { version = "4.0.26", features = ["derive", "cargo", "unicode"] }
+clap_complete = "4.0.5"
 dirs-next = "2.0.0"
-dunce = "1.0.2"
-gethostname = "0.2.3"
-# Addresses https://github.com/starship/starship/issues/4251
-git-features = { version = "0.22.6", features = ["fs-walkdir-single-threaded"] }
-git-repository = "0.24.0"
-indexmap = { version = "1.9.1", features = ["serde"] }
+dunce = "1.0.3"
+gethostname = "0.4.0"
+git-features = { version = "0.23.1", optional = true }
+# default feature restriction addresses https://github.com/starship/starship/issues/4251
+git-repository = { version = "0.28.0", default-features = false, features = ["max-performance-safe"] }
+indexmap = { version = "1.9.2", features = ["serde"] }
 local_ipaddress = "0.1.3"
-log = { version = "0.4.16", features = ["std"] }
+log = { version = "0.4.17", features = ["std"] }
 # nofity-rust is optional (on by default) because the crate doesn't currently build for darwin with nix
 # see: https://github.com/NixOS/nixpkgs/issues/160876
 notify-rust = { version = "4.5.10", optional = true }
 nu-ansi-term = "0.46.0"
-once_cell = "1.15.0"
+once_cell = "1.16.0"
 open = "3.0.3"
+# update os module config and tests when upgrading os_info
 os_info = "3.5.1"
 path-slash = "0.2.1"
-pest = "2.3.1"
-pest_derive = "2.3.1"
-quick-xml = "0.25.0"
+pest = "2.4.1"
+pest_derive = "2.4.1"
+quick-xml = "0.26.0"
 rand = "0.8.5"
-rayon = "1.5.3"
-regex = "1.6.0"
+rayon = "1.6.0"
+regex = "1.7.0"
 rust-ini = "0.18.0"
 semver = "1.0.14"
-serde = { version = "1.0.145", features = ["derive"] }
-serde_json = "1.0.85"
+serde = { version = "1.0.147", features = ["derive"] }
+serde_json = "1.0.88"
 sha-1 = "0.10.0"
-shadow-rs = { version = "0.16.3", default-features = false }
+shadow-rs = { version = "0.17.1", default-features = false }
 # battery is optional (on by default) because the crate doesn't currently build for Termux
 # see: https://github.com/svartalf/rust-battery/issues/33
 starship-battery = { version = "0.7.9", optional = true }
 strsim = "0.10.0"
-systemstat = "=0.2.0"
-terminal_size = "0.2.1"
+systemstat = "=0.2.2"
+terminal_size = "0.2.2"
 toml = { version = "0.5.9", features = ["preserve_order"] }
-toml_edit = "0.14.4"
+toml_edit = "0.15.0"
 unicode-segmentation = "1.10.0"
 unicode-width = "0.1.10"
 urlencoding = "2.1.2"
