@@ -3,6 +3,8 @@ use std::ffi::OsStr;
 use std::path::{Path, PathBuf};
 use std::{env, io};
 
+use which::which;
+
 /* We use a two-phase init here: the first phase gives a simple command to the
 shell. This command evaluates a more complicated script using `source` and
 process substitution.
@@ -23,9 +25,11 @@ struct StarshipPath {
 }
 impl StarshipPath {
     fn init() -> io::Result<Self> {
-        Ok(Self {
-            native_path: env::current_exe()?,
-        })
+        let exe_name = option_env!("CARGO_PKG_NAME").unwrap_or("starship");
+
+        let native_path = which(exe_name).or_else(|_| env::current_exe())?;
+
+        Ok(Self { native_path })
     }
     fn str_path(&self) -> io::Result<&str> {
         let current_exe = self
@@ -44,11 +48,11 @@ impl StarshipPath {
     fn sprint_pwsh(&self) -> io::Result<String> {
         self.str_path()
             .map(|s| s.replace('\'', "''"))
-            .map(|s| format!("'{}'", s))
+            .map(|s| format!("'{s}'"))
     }
     /// Command Shell specific path escaping
     fn sprint_cmdexe(&self) -> io::Result<String> {
-        self.str_path().map(|s| format!("\"{}\"", s))
+        self.str_path().map(|s| format!("\"{s}\""))
     }
     fn sprint_posix(&self) -> io::Result<String> {
         // On non-Windows platform, return directly.
@@ -229,7 +233,7 @@ pub fn init_main(shell_name: &str) -> io::Result<()> {
 
 fn print_script(script: &str, path: &str) {
     let script = script.replace("::STARSHIP::", path);
-    print!("{}", script);
+    print!("{script}");
 }
 
 /* GENERAL INIT SCRIPT NOTES
