@@ -137,7 +137,7 @@ fn describe_rebase<'a>(repo: &'a Repo, rebase_config: &'a str) -> StateDescripti
     };
 
     let (current, total) = if let Some((c, t)) = progress {
-        (Some(format!("{}", c)), Some(format!("{}", t)))
+        (Some(format!("{c}")), Some(format!("{t}")))
     } else {
         (None, None)
     };
@@ -159,13 +159,12 @@ struct StateDescription<'a> {
 mod tests {
     use nu_ansi_term::Color;
     use std::ffi::OsStr;
-    use std::fs::OpenOptions;
-    use std::io::{self, Error, ErrorKind, Write};
+    use std::io::{self, Error, ErrorKind};
     use std::path::Path;
     use std::process::Stdio;
 
     use crate::test::ModuleRenderer;
-    use crate::utils::create_command;
+    use crate::utils::{create_command, write_file};
 
     #[test]
     fn show_nothing_on_empty_dir() -> io::Result<()> {
@@ -186,7 +185,7 @@ mod tests {
         let repo_dir = create_repo_with_conflict()?;
         let path = repo_dir.path();
 
-        run_git_cmd(&["rebase", "other-branch"], Some(path), false)?;
+        run_git_cmd(["rebase", "other-branch"], Some(path), false)?;
 
         let actual = ModuleRenderer::new("git_state").path(path).collect();
 
@@ -201,7 +200,7 @@ mod tests {
         let repo_dir = create_repo_with_conflict()?;
         let path = repo_dir.path();
 
-        run_git_cmd(&["merge", "other-branch"], Some(path), false)?;
+        run_git_cmd(["merge", "other-branch"], Some(path), false)?;
 
         let actual = ModuleRenderer::new("git_state").path(path).collect();
 
@@ -216,7 +215,7 @@ mod tests {
         let repo_dir = create_repo_with_conflict()?;
         let path = repo_dir.path();
 
-        run_git_cmd(&["cherry-pick", "other-branch"], Some(path), false)?;
+        run_git_cmd(["cherry-pick", "other-branch"], Some(path), false)?;
 
         let actual = ModuleRenderer::new("git_state").path(path).collect();
 
@@ -234,7 +233,7 @@ mod tests {
         let repo_dir = create_repo_with_conflict()?;
         let path = repo_dir.path();
 
-        run_git_cmd(&["bisect", "start"], Some(path), false)?;
+        run_git_cmd(["bisect", "start"], Some(path), false)?;
 
         let actual = ModuleRenderer::new("git_state").path(path).collect();
 
@@ -249,7 +248,7 @@ mod tests {
         let repo_dir = create_repo_with_conflict()?;
         let path = repo_dir.path();
 
-        run_git_cmd(&["revert", "--no-commit", "HEAD~1"], Some(path), false)?;
+        run_git_cmd(["revert", "--no-commit", "HEAD~1"], Some(path), false)?;
 
         let actual = ModuleRenderer::new("git_state").path(path).collect();
 
@@ -289,18 +288,9 @@ mod tests {
         let path = repo_dir.path();
         let conflicted_file = repo_dir.path().join("the_file");
 
-        let write_file = |text: &str| {
-            let mut file = OpenOptions::new()
-                .write(true)
-                .create(true)
-                .truncate(true)
-                .open(&conflicted_file)?;
-            write!(file, "{}", text)
-        };
-
         // Initialize a new git repo
         run_git_cmd(
-            &[
+            [
                 "init",
                 "--quiet",
                 path.to_str().expect("Path was not UTF-8"),
@@ -311,12 +301,12 @@ mod tests {
 
         // Set local author info
         run_git_cmd(
-            &["config", "--local", "user.email", "starship@example.com"],
+            ["config", "--local", "user.email", "starship@example.com"],
             Some(path),
             true,
         )?;
         run_git_cmd(
-            &["config", "--local", "user.name", "starship"],
+            ["config", "--local", "user.name", "starship"],
             Some(path),
             true,
         )?;
@@ -325,35 +315,35 @@ mod tests {
         // If build environment has `init.defaultBranch` global set
         // it will default to an unknown branch, so need to make & change branch
         run_git_cmd(
-            &["checkout", "-b", "master"],
+            ["checkout", "-b", "master"],
             Some(path),
             // command expected to fail if already on the expected branch
             false,
         )?;
 
         // Write a file on master and commit it
-        write_file("Version A")?;
-        run_git_cmd(&["add", "the_file"], Some(path), true)?;
+        write_file(&conflicted_file, "Version A")?;
+        run_git_cmd(["add", "the_file"], Some(path), true)?;
         run_git_cmd(
-            &["commit", "--message", "Commit A", "--no-gpg-sign"],
+            ["commit", "--message", "Commit A", "--no-gpg-sign"],
             Some(path),
             true,
         )?;
 
         // Switch to another branch, and commit a change to the file
-        run_git_cmd(&["checkout", "-b", "other-branch"], Some(path), true)?;
-        write_file("Version B")?;
+        run_git_cmd(["checkout", "-b", "other-branch"], Some(path), true)?;
+        write_file(&conflicted_file, "Version B")?;
         run_git_cmd(
-            &["commit", "--all", "--message", "Commit B", "--no-gpg-sign"],
+            ["commit", "--all", "--message", "Commit B", "--no-gpg-sign"],
             Some(path),
             true,
         )?;
 
         // Switch back to master, and commit a third change to the file
-        run_git_cmd(&["checkout", "master"], Some(path), true)?;
-        write_file("Version C")?;
+        run_git_cmd(["checkout", "master"], Some(path), true)?;
+        write_file(conflicted_file, "Version C")?;
         run_git_cmd(
-            &["commit", "--all", "--message", "Commit C", "--no-gpg-sign"],
+            ["commit", "--all", "--message", "Commit C", "--no-gpg-sign"],
             Some(path),
             true,
         )?;
