@@ -109,11 +109,11 @@ impl<'a> GitDiff<'a> {
 
 #[cfg(test)]
 mod tests {
-    use crate::utils::create_command;
+    use crate::utils::{create_command, write_file};
     use std::ffi::OsStr;
     use std::fs::OpenOptions;
     use std::io::{self, Error, ErrorKind, Write};
-    use std::path::{Path, PathBuf};
+    use std::path::Path;
     use std::process::Stdio;
 
     use nu_ansi_term::Color;
@@ -139,7 +139,7 @@ mod tests {
         let path = repo_dir.path();
 
         let the_file = path.join("the_file");
-        let mut the_file = OpenOptions::new().append(true).open(&the_file)?;
+        let mut the_file = OpenOptions::new().append(true).open(the_file)?;
         writeln!(the_file, "Added line")?;
         the_file.sync_all()?;
 
@@ -157,7 +157,7 @@ mod tests {
         let path = repo_dir.path();
 
         let file_path = path.join("the_file");
-        write_file(file_path, "First Line\nSecond Line")?;
+        write_file(file_path, "First Line\nSecond Line\n")?;
 
         let actual = render_metrics(path);
 
@@ -173,7 +173,7 @@ mod tests {
         let path = repo_dir.path();
 
         let file_path = path.join("the_file");
-        write_file(file_path, "\nSecond Line\n\nModified\nAdded")?;
+        write_file(file_path, "\nSecond Line\n\nModified\nAdded\n")?;
 
         let actual = render_metrics(path);
 
@@ -205,7 +205,7 @@ mod tests {
         let path = repo_dir.path();
 
         let the_file = path.join("the_file");
-        let mut the_file = OpenOptions::new().append(true).open(&the_file)?;
+        let mut the_file = OpenOptions::new().append(true).open(the_file)?;
         writeln!(the_file, "Added line")?;
         the_file.sync_all()?;
 
@@ -263,16 +263,6 @@ mod tests {
         }
     }
 
-    fn write_file(file: PathBuf, text: &str) -> io::Result<()> {
-        let mut file = OpenOptions::new()
-            .write(true)
-            .create(true)
-            .truncate(true)
-            .open(&file)?;
-        writeln!(file, "{}", text)?;
-        file.sync_all()
-    }
-
     fn create_repo_with_commit() -> io::Result<tempfile::TempDir> {
         let repo_dir = tempfile::tempdir()?;
         let path = repo_dir.path();
@@ -280,7 +270,7 @@ mod tests {
 
         // Initialize a new git repo
         run_git_cmd(
-            &[
+            [
                 "init",
                 "--quiet",
                 path.to_str().expect("Path was not UTF-8"),
@@ -291,12 +281,12 @@ mod tests {
 
         // Set local author info
         run_git_cmd(
-            &["config", "--local", "user.email", "starship@example.com"],
+            ["config", "--local", "user.email", "starship@example.com"],
             Some(path),
             true,
         )?;
         run_git_cmd(
-            &["config", "--local", "user.name", "starship"],
+            ["config", "--local", "user.name", "starship"],
             Some(path),
             true,
         )?;
@@ -305,17 +295,17 @@ mod tests {
         // If build environment has `init.defaultBranch` global set
         // it will default to an unknown branch, so need to make & change branch
         run_git_cmd(
-            &["checkout", "-b", "master"],
+            ["checkout", "-b", "master"],
             Some(path),
             // command expected to fail if already on the expected branch
             false,
         )?;
 
         // Write a file on master and commit it
-        write_file(file, "First Line\nSecond Line\nThird Line")?;
-        run_git_cmd(&["add", "the_file"], Some(path), true)?;
+        write_file(file, "First Line\nSecond Line\nThird Line\n")?;
+        run_git_cmd(["add", "the_file"], Some(path), true)?;
         run_git_cmd(
-            &["commit", "--message", "Commit A", "--no-gpg-sign"],
+            ["commit", "--message", "Commit A", "--no-gpg-sign"],
             Some(path),
             true,
         )?;
